LPD-75173 Rename package to follow patterns
