+++ modules/dxp/apps/ai-hub/ai-hub-api/bnd.bnd
@@ -3,4 +3,4 @@ Bundle-SymbolicName: com.liferay.ai.hub.api
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.ai.hub.configuration,\
-	com.liferay.ai.hub.util
+	com.liferay.ai.hub.security
