+++ modules/dxp/apps/ai-hub/ai-hub-api/build.gradle
@@ -1,7 +1,7 @@
 dependencies {
 	compileOnly group: "com.liferay", name: "biz.aQute.bnd.annotation", version: "4.2.0.LIFERAY-PATCHED-2"
-	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "com.nimbusds", name: "nimbus-jose-jwt", version: "10.4"
+	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 	compileOnly project(":core:petra:petra-string")
 }
