+++ modules/dxp/apps/ai-hub/ai-hub-api/src/main/resources/com/liferay/ai/hub/security/packageinfo
