+++ modules/dxp/apps/ai-hub/ai-hub-api/src/test/java/com/liferay/ai/hub/security/JWTTokenUtilTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.ai.hub.util;
+package com.liferay.ai.hub.security;
 
 import com.liferay.portal.kernel.security.SecureRandomUtil;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
